feat(components): add footer component
